requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.6.3
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
Werkzeug==3.0.1
